==> notebooks/04_electric_disturbance_events.ipynb <==
# Electric Disturbance Events

## Install Dependencies

%%capture
%pip install pandas==1.3.5
%pip install numpy==1.21.6

## Import Modules

import os
import pandas as pd
import numpy as np

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
pd.options.mode.chained_assignment = None

## Mount Drive

import os

team_name = 'capstone-power-grid-protagonists'
colab_path = f'/content/drive/Shareddrives/{team_name}/project'
studiolab_path = f'/home/studio-lab-user/sagemaker-studiolab-notebooks/{team_name}'

try:
    # Try to mount Google Drive and set project path
    from google.colab import drive
    drive.flush_and_unmount()
    drive.mount('/content/drive')
    print('')

    root_path = colab_path
    os.chdir(root_path)

except:
    try:
        # Try to set AWS SageMaker Studio Lab project path
        root_path = studiolab_path
        os.chdir(root_path)
    
    except:
        # Set current working directory as root path
        root_path = os.getcwd()
        os.chdir(root_path)

        # If the current folder is 'notebooks', move up one level
        if root_path.endswith('/notebooks'):
            root_path = '/'.join(root_path.split('/')[:-1])
            os.chdir(root_path)
        
print('Current working directory is:')
print(os.getcwd())

## Import Data

# Columns to exclude in case notebook is being re-run
exclude = ['weight', 'outage', 'disturbance_prob', 'outage_prob', 'potential_population_affected', 'ba_affected']

# Processed data
cols = list(pd.read_csv('data/processed/hifld/power_plants.csv', nrows =1))
plants = pd.read_csv('data/processed/hifld/power_plants.csv',
                     low_memory=False,
                     usecols=[i for i in cols if i not in exclude])

cols = list(pd.read_csv('data/processed/hifld/substations.csv', nrows =1))
subs = pd.read_csv('data/processed/hifld/substations.csv',
                   low_memory=False,
                   usecols=[i for i in cols if i not in exclude])

cols = list(pd.read_csv('data/processed/eia/balancing_authorities.csv', nrows =1))
bas = pd.read_csv('data/processed/eia/balancing_authorities.csv',
                  low_memory=False,
                  usecols=[i for i in cols if i not in exclude])

cols = list(pd.read_csv('data/processed/doe/disturbances.csv', nrows =1))
disturb = pd.read_csv('data/processed/doe/disturbances.csv',
                      low_memory=False,
                      usecols=[i for i in cols if i not in exclude],
                      dtype={'event_type': 'category'},
                      parse_dates=['date_start', 'date_end'])

egrid = pd.read_csv('data/processed/epa/egrid.csv',
                    low_memory=False,
                    usecols=['orispl', 'fipscnty','pstatabb','bacode'])
egrid.dropna(subset=['bacode'], inplace=True)
egrid.drop_duplicates(inplace=True)

# Import states and abbreviations
states = pd.read_csv('data/raw/doe/states.csv',
                     header=None,
                     names=['name', 'abbr'])
states = dict(states.values)

# Import census data
census = pd.read_csv('data/raw/doe/census.csv',
                     usecols=['SUMLEV', 'STATE', 'COUNTY', 'STNAME','CTYNAME','POPESTIMATE2021'],
                     encoding='latin-1')
census.columns = census.columns.str.lower()
census['stname'] = census['stname'].apply(lambda x: states[x] if x in states.keys() else np.NaN)
census['county'] = census['county'].astype(str).str.pad(width=3, side='left', fillchar='0')
census['county'] = census['state'].astype(str) + census['county']
census['county'] = census['county'].astype('int64')

## Analysis

# Assign a lower weight to those disturbances where there is not county specified and only a state
disturb['weight'] = 1
disturb.loc[disturb['county_affected'].isna(), 'weight'] = .25

# Assign a boolean value as to whether there was an actual outage
disturb['outage'] = 0
disturb.loc[(disturb['demand_loss_mw'] > 0) | (disturb['num_customers_affected'] > 0), 'outage'] = 1

#### Population

# Merge disturbances, power plants, and substations, with census population data
disturb = disturb.merge(census[census['sumlev']==50][['stname', 'ctyname', 'popestimate2021']],
                        how='left',
                        left_on=['state_affected', 'county_affected'],
                        right_on=['stname', 'ctyname']) \
                 .rename(columns={'popestimate2021':'potential_population_affected'}) \
                 .drop(columns=['stname', 'ctyname'])

plants = plants.merge(census[['county', 'popestimate2021']],
                      how='left',
                      left_on=['countyfips'],
                      right_on=['county'],
                      suffixes=[None, '_census']) \
               .rename(columns={'popestimate2021':'potential_population_affected'}) \
               .drop(columns=['county_census'])

subs = subs.merge(census[['county', 'popestimate2021']],
                  how='left',
                  left_on=['countyfips'],
                  right_on=['county'],
                  suffixes=[None, '_census']) \
            .rename(columns={'popestimate2021':'potential_population_affected'}) \
            .drop(columns=['county_census'])

disturb = disturb.merge(census[census['sumlev']==40][['stname', 'popestimate2021']],
                        how='left',
                        left_on='state_affected',
                        right_on='stname') \
                 .rename(columns={'popestimate2021':'state_population'}) \
                 .drop(columns=['stname'])

disturb['potential_population_affected'].fillna(disturb['state_population']*.25, inplace=True)
disturb.drop(columns=['state_population'], inplace=True)

#### Balancing Authorities

orispl_count_by_ba = egrid.groupby(['fipscnty','bacode']).orispl.count().reset_index()
orispl_count_by_ba = orispl_count_by_ba.sort_values('orispl').drop_duplicates('fipscnty', keep='last')

# Merge the balancing authority that is most likely affected
disturb = disturb.merge(orispl_count_by_ba[['bacode','fipscnty']],
                        how='left',
                        left_on='county_affected_fips',
                        right_on='fipscnty')

disturb.rename(columns={'bacode': 'ba_affected'}, inplace=True)

# If a balancing authority was not found by county, match by state
orispl_count_by_ba = egrid.groupby(['pstatabb','bacode']).orispl.count().reset_index()
orispl_count_by_ba = orispl_count_by_ba.sort_values('orispl').drop_duplicates('pstatabb', keep='last')
state_majority_ba_dict = dict(zip(orispl_count_by_ba.pstatabb, orispl_count_by_ba.bacode))
disturb['ba_affected_by_state'] = disturb['state_affected'].apply(lambda x: state_majority_ba_dict[x])

disturb['ba_affected'].fillna(disturb['ba_affected_by_state'], inplace=True)

disturb.reset_index(drop=True, inplace=True)

# Calculate probability of a disturbance, whether there was an actual outage or not
bas = bas.merge(disturb.groupby('ba_affected')[['weight', 'potential_population_affected']].sum(),
                how='left',
                left_on='ba_code',
                right_on='ba_affected') \
         .rename(columns={'weight': 'disturbance_count'})

bas['disturbance_prob'] = bas['disturbance_count']/len(disturb['event_id'].unique())
bas['disturbance_prob'].fillna(0, inplace=True)

bas.drop(columns=['disturbance_count'], inplace=True)

# Calculate probability of an actual outage
bas = bas.merge(disturb[disturb['outage'] == 1].groupby('ba_affected')['weight'].sum(),
                how='left',
                left_on='ba_code',
                right_on='ba_affected') \
         .rename(columns={'weight': 'outage_count'})

bas['outage_prob'] = bas['outage_count']/len(disturb['event_id'].unique())
bas['outage_prob'].fillna(0, inplace=True)

bas.drop(columns=['outage_count'], inplace=True)

#### Power Plants

# Calculate probability of a disturbance, whether there was an actual outage or not
plants = plants.merge(disturb.groupby('county_affected_fips')['weight'].sum(),
                      how='left',
                      left_on='countyfips',
                      right_on='county_affected_fips') \
               .rename(columns={'weight': 'disturbance_count'})

plants = plants.merge(disturb.groupby('state_affected')['weight'].sum(),
                      how='left',
                      left_on='state',
                      right_on='state_affected',
                      suffixes=[None, '_state']) \
               .rename(columns={'weight': 'disturbance_count_state'})

# If county data was not available, fill in data based on the state which has a lower weight
plants['disturbance_count'].fillna(plants['disturbance_count_state'], inplace=True)

plants['disturbance_prob'] = plants['disturbance_count']/len(disturb['event_id'].unique())
plants['disturbance_prob'].fillna(0, inplace=True)

plants.drop(columns=['disturbance_count', 'disturbance_count_state'], inplace=True)

# Calculate probability of an actual outage
plants = plants.merge(disturb[disturb['outage'] == 1].groupby('county_affected_fips')['weight'].sum(),
                      how='left',
                      left_on='countyfips',
                      right_on='county_affected_fips') \
               .rename(columns={'weight': 'outage_count'})

plants = plants.merge(disturb[disturb['outage'] == 1].groupby('state_affected')['weight'].sum(),
                      how='left',
                      left_on='state',
                      right_on='state_affected') \
               .rename(columns={'weight': 'outage_count_state'})

plants['outage_count'].fillna(plants['outage_count_state'], inplace=True)

plants['outage_prob'] = plants['outage_count']/len(disturb['event_id'].unique())
plants['outage_prob'].fillna(0, inplace=True)

plants.drop(columns=['outage_count', 'outage_count_state'], inplace=True)

#### Substations

# Calculate probability of a disturbance, whether there was an actual outage or not
subs = subs.merge(disturb.groupby('county_affected_fips')['weight'].sum(),
                      how='left',
                      left_on='countyfips',
                      right_on='county_affected_fips') \
               .rename(columns={'weight': 'disturbance_count'})

subs = subs.merge(disturb.groupby('state_affected')['weight'].sum(),
                      how='left',
                      left_on='state',
                      right_on='state_affected',
                      suffixes=[None, '_state']) \
               .rename(columns={'weight': 'disturbance_count_state'})

# If county data was not available, fill in data based on the state which has a lower weight
subs['disturbance_count'].fillna(subs['disturbance_count_state'], inplace=True)

subs['disturbance_prob'] = subs['disturbance_count']/len(disturb['event_id'].unique())
subs['disturbance_prob'].fillna(0, inplace=True)

subs.drop(columns=['disturbance_count', 'disturbance_count_state'], inplace=True)

# Calculate probability of an actual outage
subs = subs.merge(disturb[disturb['outage'] == 1].groupby('county_affected_fips')['weight'].sum(),
                      how='left',
                      left_on='countyfips',
                      right_on='county_affected_fips') \
               .rename(columns={'weight': 'outage_count'})

subs = subs.merge(disturb[disturb['outage'] == 1].groupby('state_affected')['weight'].sum(),
                      how='left',
                      left_on='state',
                      right_on='state_affected') \
               .rename(columns={'weight': 'outage_count_state'})

subs['outage_count'].fillna(subs['outage_count_state'], inplace=True)

subs['outage_prob'] = subs['outage_count']/len(disturb['event_id'].unique())
subs['outage_prob'].fillna(0, inplace=True)

subs.drop(columns=['outage_count', 'outage_count_state'], inplace=True)

## Enriched Output

plants.sample(3)

subs.sample(3)

bas.sample(3)

# Drop and reorder columns
disturb = disturb[['event_id',
                   'date_start',
                   'date_end',
                   'state_affected',
                   'county_affected',
                   'county_affected_fips',
                   'ba_affected',
                   'nerc_region',
                   'event_type',
                   'demand_loss_mw',
                   'num_customers_affected',
                   'potential_population_affected']]

disturb.sample(3)

plants.to_csv('data/processed/hifld/power_plants.csv', index=False)
subs.to_csv('data/processed/hifld/substations.csv', index=False)
bas.to_csv('data/processed/eia/balancing_authorities.csv', index=False)
disturb.to_csv('data/processed/doe/disturbances.csv', index=False)
