==> notebooks/06_risk_analysis.ipynb <==
# Risk Analysis

## Install Dependencies

%%capture
%pip install pandas==1.3.5
%pip install numpy==1.21.6

## Import Modules

import os
import pandas as pd
import numpy as np

## Mount Drive

import os

team_name = 'capstone-power-grid-protagonists'
colab_path = f'/content/drive/Shareddrives/{team_name}/project'
studiolab_path = f'/home/studio-lab-user/sagemaker-studiolab-notebooks/{team_name}'

try:
    # Try to mount Google Drive and set project path
    from google.colab import drive
    drive.flush_and_unmount()
    drive.mount('/content/drive')
    print('')

    root_path = colab_path
    os.chdir(root_path)

except:
    try:
        # Try to set AWS SageMaker Studio Lab project path
        root_path = studiolab_path
        os.chdir(root_path)
    
    except:
        # Set current working directory as root path
        root_path = os.getcwd()
        os.chdir(root_path)

        # If the current folder is 'notebooks', move up one level
        if root_path.endswith('/notebooks'):
            root_path = '/'.join(root_path.split('/')[:-1])
            os.chdir(root_path)
        
print('Current working directory is:')
print(os.getcwd())

## Import Data

# Processed data
plants = pd.read_csv('data/processed/hifld/power_plants.csv', low_memory=False)
subs = pd.read_csv('data/processed/hifld/substations.csv', low_memory=False)
bas = pd.read_csv('data/processed/eia/balancing_authorities.csv', low_memory=False)

## Analysis

# This function will return the dataframe with defined columns ranked and a final risk score
def risk_score(df,
               cols=['deg_cent', 'bet_cent', 'clust_coef', 'disturbance_prob', 'outage_prob', 'potential_population_affected'],
               ascending=[True, True, False, True, True, True]):

    rank_cols = []

    for idx, col in enumerate(cols):
        if col in df.columns.tolist():
            rank_col = col + '_rank'
            df[rank_col] = df[col].rank(na_option='bottom', pct=True, ascending=ascending[idx])
            rank_cols.append(rank_col)

    df['risk_score'] = df[rank_cols].mean(axis=1).round(3)
    
    return df.sort_values('risk_score', ascending=False)

plants = risk_score(plants)
subs = risk_score(subs)
bas = risk_score(bas)

subs[['sub_code', 'name', 'city', 'state', 'deg_cent', 'bet_cent', 'clust_coef', 'disturbance_prob', 'outage_prob', 'potential_population_affected', 'risk_score']].head(25)

bas[['ba_code', 'ba_name', 'region_country_name', 'deg_cent', 'bet_cent', 'clust_coef', 'disturbance_prob', 'outage_prob', 'potential_population_affected', 'risk_score']].head(10)

subs.groupby('state')[['state', 'deg_cent', 'bet_cent', 'clust_coef', 'disturbance_prob', 'outage_prob', 'potential_population_affected', 'risk_score']].mean().sort_values(by='risk_score', ascending=False).head(10)

bas.groupby('region_country_name')[['region_country_name', 'deg_cent', 'bet_cent', 'clust_coef', 'disturbance_prob', 'outage_prob', 'potential_population_affected', 'risk_score']].mean().sort_values(by='risk_score', ascending=False).head(10)
